only delete reports table if table exists

+++ db/migrate/20150811142301_remove_reports_table.rb
@@ -1,5 +1,5 @@
 class RemoveReportsTable < ActiveRecord::Migration
   def change
-    drop_table :reports
+    drop_table :reports if ActiveRecord::Base.connection.table_exists? 'reports'
   end
 end
